requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Config
pyyaml>=6.0
python-dotenv>=1.0.0

# Async I/O
aiofiles>=23.0.0
websockets>=12.0.0

# Logging
python-json-logger>=2.0.0

# System Monitoring
psutil>=5.9.0
